Send payments to blinded routes
Add support for paying blinded routes. We introduce a new type of recipient
for those payments, that uses blinded graph edges and creates onion payloads
accordingly. There is a subtlety in the case where we're the introduction
of the blinded route: when that happens we need to decrypt the first payload
to figure out where to send the payment.

When we receive a failure from a blinded route, we simply ignore it in
retries: we don't know what caused the issue so we assume it's permanent,
which makes sense in most cases since we cannot change the relaying
parameters (fees and expiry delta are chosen by the recipient).