introduce MetaCommitments
We use it in channel states and keep changes minimal by defining:
```scala
  def commitments: Commitments = metaCommitments.main
```

Note that, despite storing multiple commitments in `MetaCommitments`, we
only support updating the main commitments. This is because the only
case where we currently can have multiple commitments is when we rbf a
funding tx, where no updates to the commitments (htlcs or fee) are
possible.

We also add a custom data field to MetaCommitments, that is unused.