store funding tx in Commitments
There is no functional change, we just move the funding tx from the
`DATA_*` level to the `Commitments` level.

With rbf support for dual funding, we started to have to support several commitments in parallel, and had to track their corresponding funding txes next to them until they were confirmed, which was quite awkward. This trend is going to amplify with the support of splices, because now we will also have several commitments in parallel also in `NORMAL` state.

There is a chicken-and-egg problem for the first commitments, but we can easily work around it, and overall it significantly simplifies channel state classes.