support updating multiple commitments in parallel
The first version of `MetaCommitments` defined a `main` commitments, and
a separate list of `rbfed` commitments.

In this commit, we replace both of those with a single list of
commitments. It is a semantic change: now all commitments are active in
parallel, and can be updated.

This is why we also define a new tlv for the `CommitSig` message, which
contains signatures for other commitments.

Even if it could still make sense to have a `main` commitments (the one
which funding tx is currently confirmed and unspent as far as we know),
it is much easier from an implementation point of view to manage a list
of commitments, which we can update (add/sig/rev/fee/etc.) in one pass
with standard collection methods.

For convenience, we keep a `main` commitments, but it now points to the
first commitments in the list, which is the most recent (another
semantic change).