Implement correct ordering of `tx_signatures`
According to the dual funding specification, the peer that has contributed
the smaller amount of `tx_add_input` must sign first. We incorrectly used
the contributions to the funding output instead of the sum of the inputs
values.

This requirement guarantees that there can be no deadlocks when nodes
are batching multiple interactive-tx sessions.