Implement correct ordering of `tx_signatures` (#2501)
* Implement correct ordering of `tx_signatures`

According to the dual funding specification, the peer that has contributed
the smaller amount of `tx_add_input` must sign first. We incorrectly used
the contributions to the funding output instead of the sum of the inputs
values.

This requirement guarantees that there can be no deadlocks when nodes
are batching multiple interactive-tx sessions.

* Send error instead tx_abort in interactive-tx

When building the first version of a dual-funded transaction, we should
use `error` instead of `tx_abort` if we encounter a failure. It makes it
more explicit to our peer that the channel should be closed at that point,
whereas `tx_abort` can be used in RBF attempts to abort a secondary
interactive-tx sessions without closing the channel.