Fix ZeroConfAliasIntegrationSpec
A test was added to verify that when building a route to self, we use the
correct scid-alias when available (remote alias instead of local).

This test sometimes failed because Carol hadn't received Alice's channel
update and couldn't build the Alice->Bob hop. The only thing that we meant
to test was the Bob->Carol hop, so we can simplify that and remove the
flakiness,