Remove redundant log messages
Reduce the amount of logs generated by default by moving logs that are
redundant or only used for targeted debugging to `debug` level or removing
them entirely.

I focused on logs that generate some volume and that I never personnally
use for day-to-day log browsing, but let me know if some of these logs
are important for you and it would be annoying to have to turn on `debug`
level for their respective component.