split Commitments between params/common/commitment
We split `Commitments` between:
- `Params`: static parameters
- `Common`: dynamic data shared across all commitments (e.g. current
changes, revocation hashes)
- `Commitment`: local+remote commit associated to one funding tx

This allow us to not store duplicate info in `MetaCommitments`.

To keep changes minimal we leave the `Commitments` class as-is for now.
It is a transient class built on-the-fly to update the commitment, but
is not persisted.

The `sendX`/`receiveX` methods in `MetaCommitments` are a bit cumbersome
because we keep converting back and forth between `Commitment` and
`Commitments`, but the trade-off seems worth it.