Always lock utxos when funding transactions
We must *always* lock inputs selected for funding, otherwise locking
wouldn't work at all, as the following scenario highlights:

- we fund a transaction for which we don't lock utxos
- we fund another unrelated transaction for which we lock utxos
 - the second transaction ends up using the same utxos as the first one
 - but the first transaction confirms, invalidating the second one

This would break the assumptions of the second transaction: its inputs
are locked,  so it doesn't expect to potentially be double-spent.

It's particularly dangerous if the node supports zero-conf, as this could
happen to a zero-conf transaction and completely break the channel.