factor handling of unknown spending tx
When the funding tx is spent by an unknown tx, we move to the `ERR_INFORMATION_LEAK` state.

But we currently have two different behaviors:
- early states `WAIT_FOR_*READY/CONFIRMED` published the local commit and sent an error
- other states did nothing

Publishing the local commit might sound like the appropriate thing to do, at first glance, but then we should go to `CLOSING` to handle the rest of the process. Maybe we should wait until an unknown spending tx is _confirmed_ before moving to `ERR_INFORMATION_LEAK`.

Anyway this being a corner case, for now I just removed the special handler for the early states.

I also added tests, and noticed that we don't test all funding-spent cases in all states. The fact that we now have factored the handling of those makes systematic testing less critical though.