Factor funding tx acceptance (#2557)
* homogenize WatchFundingConfirmedTriggered handlers

The handling is mostly the same in connected/disconnected states. This
does some refactoring to make them as similar as possible.

NB: we don't make a difference between previous/latest commitment being
confirmed.

* properly factor the pruning of commitments

This is a first glimpse of what will happen when we prune commitments
that have been overriden by a splice.

* prevent duplicate logs when funding tx confirms

At restart, we always put a `WatchFundingConfirmed` on all funding txs,
because this is how we detect force close in state disconnected (we only
put a `WatchSpent` on the funding tx output _after_ we detect that the
funding tx itself is confirmed).

With current code, this would result in a log indicating that the
funding tx has been confirmed at restart, for each channel in
`WAIT_FOR_DUAL_FUNDING_CONFIRMED` state. It can be quite spammy.

Instead, we now only log two cases:
- when we are disconnected, and there were alternative funding txs that
we discard;
- when we are connected (this can only happen once because we will
then transition to the next state).

* factor code for accepting the funding tx

There are 4 code paths:
- single/dual funding
- zero/non-zero conf

Besides the factorization, note that in the dual funding + non-zero conf case, we reverse the calling order between `pruneCommitments` and `acceptDualFundingTx`. This prepares a future change where we store the funding tx inside `Commitments`: it makes more sense then to first update the status of the funding tx in the corresponding commitment, and then prune the remaining commitments.