read channel db only once at restart
Fixes #2478.

I tried to keep the changes minimal, but had to move the _read_ from `Setup` to `Setup.bootstrap`, otherwise we would have kept a reference of the initial channels list forever.

Checking the db now happens a bit further in the startup procedure, but it is still blocking, and before the actors are initiated.