Never lock utxos when funding
We previously locked utxos when funding transactions to avoid
accidentally double-spending ourselves if an unrelated transaction was
funded concurrently before the first one was published.

It didn't fully protect us against double-spending ourselves though. If
bitcoind was restarted while eclair was running, it would clear existing
locks and remove those guarantees. But that was an acceptable edge case.

However, dual funding and liquidity ads introduce griefing attacks if we
lock our utxos: an attacker can waste our liquidity in funding attempts
that will never be published, as explained here:
https://github.com/lightning/bolts/pull/851#discussion_r997537630

The easiest way to get rid of that griefing attack is to never lock our
inputs: this ensures that honest users will double-spend the inputs used
for sessions with malicious users, at no cost to us. It also greatly
simplifies our code because we don't need to worry about unlocking utxos.

Note that once a transaction has been successfully published once, we
are then guaranteed that bitcoind will never accidentally double-spend
it: we have to call `abandontransaction` to tell bitcoind to forget that
transaction and free its inputs.

There is one drawback with this approach: if we have concurrent sessions
with honest peers, there is a risk that we will double-spend ourselves
and only one of the sessions will succeed. This trade-off is acceptable
because:

- the time window where such race condition may happen is small: honest
  peers will quickly complete the funding flow and publish the signed
  transaction, which protects against accidental double-spends
- funding on-chain transactions is an infrequent operation: block size
  is limited, so we have to rate-limit our funding attempts anyway
- to be perfectly safe, all funding code must handle the possibility of
  being double-spent before the transaction is published: the main
  danger here was zero-conf channels, but it has been fixed in #2558.
- all existing funding scenarios already handle the case where the
  transaction cannot be published and detect double-spends
- when a funding attempt fails, it can simply be restarted (e.g. LSP
  operations that open channels on-the-fly)

It's also important to note that bitcoind funding locks have to either
be used all the time or never. We cannot use them in some scenarios and
not others, because it wouldn't provide any meaningful guarantees for
the scenarios that lock inputs, as shown by the following sample flow:

- fund a transaction for which we don't lock utxos
- fund another unrelated transaction for which we lock utxos
- the second transaction ends up using the same utxos as the first one,
  since they were unlocked and thus ready to be used
- but the first transaction confirms, invalidating the second one, which
  incorrectly thought it was safe from double-spends because it used locks