Allow keysend payments without a payment secret
Apparently lnd and cln don't send a payment secret when using keysend,
even when we advertize the feature as mandatory. This is ok though, we
don't actually need a payment secret to protect keysend payments as long
as the sender doesn't use MPP (which is their responsibility, not ours).

Since our codebase requires payment secrets to index payments, we simply
use the preimage as secret when receiving keysend payments.