Move htlc helper functions
This commit contains small changes because the next remote commit is
an option in `Commitment` instead of accessed as a `Left` in `Commitments.

The code is actually more readable this way, and highlights the fact that
this isn't something that belongs to individual commitments, but is rather
shared between all of them.