Remove default access to latest Commitments
This commit looks big, but it only contains very mechanical refactorings.

We remove access to a default `Commitments` object in
`PersistentChannelData` and explicitly use the latest commitment when
necessary (which we rename `latest` instead of `main`).

All code that doesn't actually need a specific commitment now uses the
shared parts (`Params` and `Common`). This will make it easier to identify
what code needs to be updated when implementing splicing. It also makes
future changes on the data model of `MetaCommitments` less impactful
on the codebase, which will help reviewers focus on the data model itself.