Remove unnecessary dust limit hack in interactive-tx (#2594)
We previously inflated the funding amount to always exceed the dust limit,
otherwise bitcoind would reject the funding attempt. It complicated the
code for a scenario that doesn't make sense in practice, so it's better
to just let the funding attempt fail in that case.

We also randomize the order of inputs and outputs, which is a common
privacy best practice.