Remove Common class from MetaCommitments
There are two possible strategies for MetaCommitments:

1. Move the duplication between every Commitment in Common
2. Access duplicated fields through the first Commitment

I initially wanted to go with solution 1., but it makes commitments really
hard to work with. The main reason is that the `CommitmentSpec` abstraction
is a very good abstraction to work with when updating commitments,
and it requires keeping the htlcs inside every commitment.

As long as we update commitments in a loop, there is no risk of common
values being desynchronized. Since they contain mostly pointers to shared
data, the memory overhead is negligible, as long as we make sure we don't
duplicate the data when serializing to disk (which will be addressed in a
future commit).

We're thus choosing solution 2., in which case it makes sense to completely
get rid of the Common class, because it otherwise adds another layer of
unnecessary duplication.