Small receiveRevocation refactoring
This one doesn't make sense to apply to all commitments.
We have a slightly weird dependency on the first commitment to fetch htlcs
and compute dust exposure: the dust exposure part could be refactored to
remove usage of `CommitmentSpec`, which would make that code slightly
better. This will be done in a separate PR.