Serialize htlcs only once
This change is potentially risky and requires careful review.
Commitments may contain a lot of htlcs, which may use a lot of memory.
It would be a waste to duplicate this information for each active
commitment, so we serialize htlcs separately and rebuild the state when
deserializing.

We also only keep the main htlc fields during json serialization to avoid
performance issues when writing json fields to the DB. There was no good
reason to serialize everything as we previously did.