Rename and document DummyBolt12Invoice
Dummy is usually used for test or invalid data, whereas here we're just
storing a minimal version of the Bolt 12 invoice, so we rename it
MinimalBolt12Invoice.

We also add an explicit constructor for it that forces callers to provide
the data we expect to store.

Note that we remove the `features` field: using `nodeParams.features`
could be incorrect if the invoice_request's features don't include some
of our node's optional features. This is unused now anyway, but more
future-proof this way.