Implement WaitForDualFundingSigned state
Add a new state to the dual-funded channel state machine where we exchange
our signatures: we exit that state once we've send our `tx_signatures`.

This is a state that will be persisted, but doesn't have a `Commitments`
because we haven't yet received our peer's `commit_sig`. We thus need to
split the persistent channel states into two distinct traits, one that has
commitments and one that doesn't. The impact on the codebase isn't as
complex as I originally thought.