Persist interactive-tx signing session
We need to persist state when we send our `commit_sig`. This is true for
the initial funding attempt and for RBF attempts.

I don't know if we should remove the HTLC duplication between the local
and remote commit: I haven't bothered with it, but maybe it's worth the
extra effort?