Simplify how Eclair manages Bitcoin Core's private keys
We don't even need to use Bitcoin Core's external signer interface, instead we can simply create an empty
wallet and import descriptors generated by Eclair.
This is functionnaly equivalent to what we had (we disabled tx singing from Bitcoin Core) and much simpler.
This is actually a small change and the signing worflow remains the same.