Update RBF contributions to allow negative amounts
When RBF-ing a splice, we will need to be consistent with the
`splice_init` and `splice_ack` messages and specify the funds we're
adding or removing from the channel instead of the absolute value.

We thus change the `tx_init_rbf` and `tx_ack_rbf` messages to use signed
amounts.