Allow negative contributions in `InteractiveTxBuilder` (#2619)
* Allow negative contributions in InteractiveTxBuilder

When splicing over an existing channel, it is simpler to specify the amount
we will add or remove from the channel, rather than the resulting amount
that takes the previous balance into account. It removes the need for
truncating msat balances and trivially carries over those balances to
the new commitment.

* Update RBF contributions to allow negative amounts

When RBF-ing a splice, we will need to be consistent with the
`splice_init` and `splice_ack` messages and specify the funds we're
adding or removing from the channel instead of the absolute value.

We thus change the `tx_init_rbf` and `tx_ack_rbf` messages to use signed
amounts.