Use channel capacity in minDepth computation
We must use the whole channel capacity to scale the number of on-chain
confirmations, because that's the amount that could be stolen from us if
the funding transaction is reorg-ed.

When using dual-funding or splicing, since we compute that amount before
going through the interactive-tx construction, we cannot know beforehand
if our peer will have inputs in the funding transaction and thus be able
to double-spend it, so we must act as if they were contributing.

We could however decide to use the result early if they send channel_ready
or splice_locked before our watch and we're the only contributor to that
funding transaction. This is already done for dual funding, and could be
added for splicing.