Remove `minDepth` from `InteractiveTxParams`
We can recompute `minDepth` based on our default config, the channel
capacity and our local features.

The only parameter that could change is our local features, which could
create issues if we enable/disable zero-conf in the middle of a funding
attempt: we may accept an RBF attempt for a transaction that we previously
treated as zero-conf, which will break the channel. But since activating
zero-conf means we have trust in our peer, and this is an unlikely
scenario, maybe this is ok?