Improvements for splice support (#2632)
* Remove funding_txid from commit_sig

And a couple of nits.

* Compute splice contribution in InteractiveTxFunder

This is where we have the rest of the relevant code (which delegates to
bitcoind if we're splicing additional funds in the channel).

We also explicitly provide the shared input, which is more future proof
when we have other types of shared inputs (e.g. taproot channels).

* Small fixes to Commitments methods

- drop obsolete commit_sigs
- fix updateFundingStatus
- add comments

* Use channel capacity in minDepth computation

We must use the whole channel capacity to scale the number of on-chain
confirmations, because that's the amount that could be stolen from us if
the funding transaction is reorg-ed.

When using dual-funding or splicing, since we compute that amount before
going through the interactive-tx construction, we cannot know beforehand
if our peer will have inputs in the funding transaction and thus be able
to double-spend it, so we must act as if they were contributing.

We could however decide to use the result early if they send channel_ready
or splice_locked before our watch and we're the only contributor to that
funding transaction. This is already done for dual funding, and could be
added for splicing.

* Various nits

- add more comments and fix a few nits in Channel.scala
- reactivate tests for unrecognized transactions confirmed

* Fix RBF attempts deactivation

And revert changes to `WaitForDualFundingConfirmedStateSpec`.

There was an issue with the following scenario:

- we create fundingTx1 (at index=42)
- we RBF it to create fundingTx2 (at index=42)
- fundingTx1 confirms
- our peer sends splice_locked for fundingTx1

The previous code from `deactivateCommitments` would select fundingTx2
and deactivate fundingTx1, which is invalid. We fix that by preserving
order among RBF attempts (even though they share the same fundingIndex).

There was another issue with the initial funding (index=0). Since there
was no explicit locking from our peer, we were keeping all RBF attempts
even after one confirmed and invalidated the other ones. We're never
exchanging `commit_sig` so we can eagerly deactivate RBF attempts,
instead of the hackish lock added when receiving `channel_ready`, where
we actually can't know which funding tx the remote is locking...